backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "talentchainpro"
version = "0.1.0"
description = "TalentChain Pro - Blockchain-based talent ecosystem with AI-powered reputation oracles"
readme = "README.md"
authors = [
    {name = "TalentChain Pro Team", email = "info@talentchainpro.io"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.5",
    "hedera-sdk-py>=2.24.0",
    "langchain>=0.0.267",
    "groq>=0.4.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.4.1",
]

[tool.setuptools]
package-dir = {"" = "backend"}

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
